0.8.11

+++ app/Http/Controllers/Learn/CardController.php
@@ -3,23 +3,27 @@
 
 namespace App\Http\Controllers\Learn;
 
+use Gate;
 use App\Helpers\Auth;
-use App\Http\Controllers\Controller;
-use App\Http\Requests\CreateCardRequest;
-use App\Http\Requests\SearchRequest;
-use Illuminate\Auth\Access\AuthorizationException;
-use Illuminate\Http\JsonResponse;
-use Scandinaver\Common\Domain\Model\Language;
-use Scandinaver\Learn\Domain\Exceptions\CardAlreadyAddedException;
 use Illuminate\Http\Request;
-use Scandinaver\Learn\Domain\Services\WordService;
-use Scandinaver\Learn\UI\Command\AddCardToAssetCommand;
-use Scandinaver\Learn\UI\Command\CreateCardCommand;
-use Scandinaver\Learn\UI\Command\DeleteCardFromAssetCommand;
-use Scandinaver\Learn\UI\Command\UpdateCardCommand;
-use Scandinaver\Learn\UI\Command\UploadCsvSentencesCommand;
-use Scandinaver\Shared\CommandBus;
 use Scandinaver\Shared\QueryBus;
+use Doctrine\DBAL\DBALException;
+use Illuminate\Http\JsonResponse;
+use Scandinaver\Shared\CommandBus;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\SearchRequest;
+use App\Http\Requests\CreateCardRequest;
+use Scandinaver\Common\Domain\Model\Language;
+use Scandinaver\Learn\Domain\Permissions\Card;
+use Scandinaver\Shared\EventBusNotFoundException;
+use Illuminate\Auth\Access\AuthorizationException;
+use Scandinaver\Learn\Domain\Services\WordService;
+use Scandinaver\Learn\UI\Command\CreateCardCommand;
+use Scandinaver\Learn\UI\Command\UpdateCardCommand;
+use Scandinaver\Learn\UI\Command\AddCardToAssetCommand;
+use Scandinaver\Learn\UI\Command\UploadCsvSentencesCommand;
+use Scandinaver\Learn\UI\Command\DeleteCardFromAssetCommand;
+use Scandinaver\Learn\Domain\Exceptions\LanguageNotFoundException;
 
 /**
  * Class CardController
@@ -31,44 +35,80 @@ class CardController extends Controller
 
     private WordService $wordService;
 
-    public function __construct(CommandBus $commandBus, QueryBus $queryBus, WordService $wordService)
-    {
+    public function __construct(
+      CommandBus $commandBus,
+      QueryBus $queryBus,
+      WordService $wordService
+    ) {
         parent::__construct($commandBus, $queryBus);
         $this->wordService = $wordService;
     }
 
     public function index()
     {
-
     }
 
     public function show()
     {
-
     }
 
+    /**
+     * @param  string  $language
+     * @param  int     $card
+     * @param  int     $asset
+     *
+     * @return JsonResponse
+     * @throws AuthorizationException
+     * @throws EventBusNotFoundException
+     */
     public function store(string $language, int $card, int $asset): JsonResponse
     {
-        try {
-            $this->commandBus->execute(new AddCardToAssetCommand(Auth::user(), $language, $card, $asset));
-        } catch (CardAlreadyAddedException $e) {
-            return response()->json($e->getMessage(), 500);
-        }
-        return response()->json(NULL, 201);
+        Gate::authorize(Card::CREATE);
+
+        return $this->execute(
+          new AddCardToAssetCommand(Auth::user(), $language, $card, $asset),
+          JsonResponse::HTTP_CREATED
+        );
     }
 
+    /**
+     * @param  int      $card
+     * @param  Request  $request
+     *
+     * @return JsonResponse
+     * @throws AuthorizationException
+     */
     public function update(int $card, Request $request): JsonResponse
     {
-        return response()->json($this->commandBus->execute(new UpdateCardCommand($card, $request->toArray())));
+        Gate::authorize(Card::UPDATE, $card);
+
+        return response()->json(
+          $this->commandBus->execute(new UpdateCardCommand($card, $request->toArray()))
+        );
     }
 
+
+    /**
+     * @param  Language           $language
+     * @param  CreateCardRequest  $request
+     *
+     * @return JsonResponse
+     * @throws AuthorizationException
+     * @throws EventBusNotFoundException
+     */
     public function create(Language $language, CreateCardRequest $request): JsonResponse
     {
-        $card = $this->commandBus->execute(new CreateCardCommand(
-            Auth::user(), $language, $request->get('word'), $request->get('translate')
-        ));
+        Gate::authorize(Card::CREATE);
 
-        return response()->json($card, 201);
+        return $this->execute(
+          new CreateCardCommand(
+            Auth::user(),
+            $language,
+            $request->get('word'),
+            $request->get('translate')
+          ),
+          JsonResponse::HTTP_CREATED
+        );
     }
 
     /**
@@ -81,24 +121,46 @@ class CardController extends Controller
      */
     public function destroy(string $language, int $card, int $asset): JsonResponse
     {
-      //  $this->authorize('delete', [$card, $asset]);
+        Gate::authorize(Card::DELETE, [$card, $asset]);
 
         $this->commandBus->execute(new DeleteCardFromAssetCommand(Auth::user(), $card, $asset));
 
-        return response()->json(NULL, 204);
+        return response()->json(null, 204);
     }
 
+    /**
+     * @param  string         $language
+     * @param  SearchRequest  $request
+     *
+     * @return JsonResponse
+     * @throws AuthorizationException
+     * @throws DBALException
+     * @throws LanguageNotFoundException
+     */
     public function search(string $language, SearchRequest $request): JsonResponse
     {
+        Gate::authorize(Card::SEARCH);
+
         $words = $this->wordService->translate($language, $request);
 
         return response()->json($words);
     }
 
-    public function uploadSentences(string $languageId, Request $request)
+    /**
+     * @param  string   $languageId
+     * @param  Request  $request
+     *
+     * @return JsonResponse
+     * @throws EventBusNotFoundException
+     */
+    public function uploadSentences(string $languageId, Request $request): JsonResponse
     {
         $file = $request->file('file');
 
-        return $this->execute(new UploadCsvSentencesCommand($languageId, $file), JsonResponse::HTTP_CREATED);
+        return $this->execute(
+          new UploadCsvSentencesCommand($languageId, $file),
+          JsonResponse::HTTP_CREATED
+        );
     }
+
 }
